move SharedAllocationDisableTrackingGuard to Kokkos_SharedAlloc, change assert to KOKKOS_ASSERT, and removex extraneous noexcepts
