+++ core/src/Kokkos_View.hpp
@@ -1919,46 +1919,6 @@ KOKKOS_INLINE_FUNCTION bool operator!=(const View<LT, LP...>& lhs,
 
 namespace Kokkos {
 namespace Impl {
-struct SharedAllocationDisableTrackingGuard {
-  SharedAllocationDisableTrackingGuard() {
-    assert(
-        (Kokkos::Impl::SharedAllocationRecord<void, void>::tracking_enabled()));
-    Kokkos::Impl::SharedAllocationRecord<void, void>::tracking_disable();
-  }
-
-  SharedAllocationDisableTrackingGuard(
-      const SharedAllocationDisableTrackingGuard&) = delete;
-  SharedAllocationDisableTrackingGuard(
-      SharedAllocationDisableTrackingGuard&&) noexcept = delete;
-
-  ~SharedAllocationDisableTrackingGuard() {
-    Kokkos::Impl::SharedAllocationRecord<void, void>::tracking_enable();
-  }
-
-  // clang-format off
-  // The old version of clang format we use is particularly egregious here
-  SharedAllocationDisableTrackingGuard& operator=(
-      const SharedAllocationDisableTrackingGuard&) = delete;
-  SharedAllocationDisableTrackingGuard& operator=(
-      SharedAllocationDisableTrackingGuard&&) noexcept = delete;
-  // clang-format on
-};
-
-template <class FunctorType, class... Args>
-inline FunctorType construct_with_shared_allocation_tracking_disabled(
-    Args&&... args) {
-  [[maybe_unused]] auto guard = SharedAllocationDisableTrackingGuard{};
-  return {std::forward<Args>(args)...};
-}
-
-} /* namespace Impl */
-} /* namespace Kokkos */
-
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-namespace Kokkos {
-namespace Impl {
 
 template <class Specialize, typename A, typename B>
 struct CommonViewValueType;
